Fix #36: Refactor flutter_demo to use Provider with feature-based organization
- Add provider package dependency
- Extract counter state into CounterProvider (ChangeNotifier)
- Break up main.dart into app.dart, counter_page.dart, and counter_provider.dart
- Organize code by features (lib/features/counter/)
- Convert MyHomePage from StatefulWidget to StatelessWidget using Consumer
- Update test to import from new app.dart location
