+++ projects/flutter_demo/lib/features/counter/counter_provider.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/foundation.dart';
+
+class CounterProvider extends ChangeNotifier {
+  int _count = 0;
+
+  int get count => _count;
+
+  void increment() {
+    _count++;
+    notifyListeners();
+  }
+}
